reset timer in power system regardless
