+++ src/main/java/net/mrscauthd/boss_tools/gui/FuelRefineryGUIGuiWindow.java
@@ -5,6 +5,7 @@ import org.lwjgl.opengl.GL11;
 
 import net.minecraftforge.fluids.capability.templates.FluidTank;
 import net.mrscauthd.boss_tools.machines.FuelRefineryBlock.CustomTileEntity;
+import net.mrscauthd.boss_tools.machines.machinetileentities.PowerSystem;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.energy.EnergyStorage;
 import net.minecraftforge.api.distmarker.Dist;
@@ -49,7 +50,7 @@ public class FuelRefineryGUIGuiWindow extends ContainerScreen<FuelRefineryGUIGui
 		this.renderHoveredTooltip(ms, mouseX, mouseY);
 
 		CustomTileEntity tileEntity = (CustomTileEntity) this.world.getTileEntity(this.blockPos);
-		EnergyStorage energyStorage = tileEntity.getEnergyStorage();
+		PowerSystem powerSystem = tileEntity.getPowerSystem();
 		FluidTank fluidTank = tileEntity.getFluidTank();
 		int fuel = fluidTank.getFluidAmount();
 
@@ -67,7 +68,7 @@ public class FuelRefineryGUIGuiWindow extends ContainerScreen<FuelRefineryGUIGui
 		// ToolTip Ende
 		// toolTipStart Energy
 		if (mouseX > guiLeft + 143 && mouseX < guiLeft + 168 && mouseY > guiTop + 20 && mouseY < guiTop + 69) {
-			this.renderTooltip(ms, new StringTextComponent(energyStorage.getEnergyStored() + " FE / " + energyStorage.getMaxEnergyStored() + " FE"), mouseX, mouseY);
+			this.renderTooltip(ms, new StringTextComponent(powerSystem.getStored() + " FE / " + powerSystem.getCapacity() + " FE"), mouseX, mouseY);
 		}
 	}
 
@@ -82,7 +83,7 @@ public class FuelRefineryGUIGuiWindow extends ContainerScreen<FuelRefineryGUIGui
 		// Energy NBT
 
 		CustomTileEntity tileEntity = (CustomTileEntity) this.world.getTileEntity(this.blockPos);
-		int energyanimation = tileEntity.getEnergyStorage().getEnergyStored();
+		int energyanimation = tileEntity.getPowerSystem().getStored();
 
 		// Fuel Tank
 		int lavaanimation = tileEntity.getFluidTank().getFluidAmount();
